pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omislisi-accounting"
version = "0.1.0"
description = "Bank and PayPal transaction analysis tool"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.0",
    "lxml>=4.9.0",
    "pandas>=2.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
oa = "omislisi_accounting.cli.main:cli"

